user_profile: optimize code
